#20 fixed UI swapi issues
