--- src/app/[lng]/portfolio/projects/swapi/about/page.tsx
@@ -1,33 +1,0 @@
-import { Metadata } from "next";
-
-export const metadata: Metadata = {
-  title: "About Swapi",
-  description: "About Swapi task",
-  keywords: ["About", "Swapi"],
-};
-
-const PageAbout = () => {
-  return (
-    <div>
-      <p>
-        This task was inspired by all the practical tasks I did regarding Swap
-        Api
-      </p>
-      <a
-        href="https://www.youtube.com/watch?v=8aGhZQkoFbQ&ab_channel=JSConf"
-        target="_blank">
-        Event Loop
-      </a>
-      <p>
-        Javascript is a single-threaded non-blocking asynchronous concurrent
-        language.
-      </p>
-      <p>
-        I have call stack an event loop a callback queue some other apis and
-        stuff.
-      </p>
-      <p>Everything there is to know about async</p>
-    </div>
-  );
-};
-export default PageAbout;
