+++ src/app/i18n/locales/en/swapi.json
@@ -1,0 +1,6 @@
+{
+  "swapi": "SWAPI",
+  "about": "About",
+  "films": "Films",
+  "portfolio": "Portfolio"
+}
