+++ src/app/i18n/locales/lt/swapi.json
@@ -1,0 +1,6 @@
+{
+  "swapi": "SWAPI",
+  "about": "Apie",
+  "films": "Filmai",
+  "portfolio": "Portfelis"
+}
