+++ src/components/layout/side-menu/side-menu.tsx
@@ -1,11 +1,11 @@
 "use client";
-import AboutIcon from "@/styles/icons/about-icon";
 import FilmsIcon from "@/styles/icons/films-icon";
-import LogoIcon from "@/styles/icons/logo-icon";
 import Link from "next/link";
 import { Fragment, ReactNode } from "react";
 import { usePathname } from "next/navigation";
 import PortfolioIcon from "@/styles/icons/portfolio-icon";
+import { useTranslation } from "@/app/i18n/client";
+import AboutIcon from "@/styles/icons/about-icon";
 
 interface IMenuItems {
   title: string;
@@ -15,20 +15,21 @@ interface IMenuItems {
 
 const SideMenu = ({ language }: { language: string }) => {
   const pathName = usePathname();
+  const { t } = useTranslation({ language, ns: "swapi" });
 
   const menuItems: IMenuItems[] = [
     {
-      title: "About",
+      title: t("about"),
       icon: <AboutIcon />,
-      link: `/${language}/portfolio/projects/swapi/about`,
+      link: `/${language}/portfolio/projects/swapi`,
     },
     {
-      title: "Films",
+      title: t("films"),
       icon: <FilmsIcon />,
       link: `/${language}/portfolio/projects/swapi/films`,
     },
     {
-      title: "Portfolio",
+      title: t("portfolio"),
       icon: <PortfolioIcon />,
       link: `/${language}/portfolio`,
     },
@@ -36,13 +37,14 @@ const SideMenu = ({ language }: { language: string }) => {
 
   return (
     <>
-      <div className="fixed top-0 left-0 z-40 w-16 h-screen overflow-hidden text-gray-400 bg-gray-900 sm:w-40">
+      <div className="fixed top-0 left-0 z-40 w-18 h-screen overflow-hidden text-gray-400 bg-gray-900 sm:w-40">
         <Link
           className="flex items-center w-full px-3 mt-3"
           // @ts-ignore
-          href="/">
-          <LogoIcon />
-          <span className="ml-2 text-sm font-bold hidden sm:block">SWAPI</span>
+          href={`/${language}/portfolio/projects/swapi`}>
+          <span className="ml-2 text-sm font-bold hidden sm:block">
+            {t("swapi")}
+          </span>
         </Link>
         <div className="w-full px-2">
           <div className="flex flex-col items-center w-full mt-3 border-t border-gray-700">
